# internal/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: MajorGTM API
  version: "1.57.1"
  description: API for MajorGTM recorded updates and async video messaging.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://rec.majorgtm.com
    description: Production
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
        name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    AuthResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
...
    RegisterResponse:
      type: object
      required: [message, requiresEmailConfirmation]
      properties:
        message:
          type: string
        requiresEmailConfirmation:
          type: boolean
          description: |-
            When false, the account was created with `email_verified=true` because no
            email backend (Listmonk / SMTP / sendmail) is configured. The client should
            redirect to `/login` instead of `/check-email`.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |-
        Rate limited to 0.5 requests per second with a burst of 5. Returns 403 when registration is disabled via `REGISTRATION_ENABLED=false`.

        When no email backend (Listmonk / SMTP / sendmail) is configured the account is created with `email_verified=true` and the response sets `requiresEmailConfirmation=false` so the client can route the user straight to login instead of the check-your-email page.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created. The body's `requiresEmailConfirmation` field tells the client whether a confirmation email was sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Registration is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      description: Rate limited to 0.5 requests per second with a burst of 5.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: refresh_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 'Email not verified. Error body contains {"error": "email_not_verified"}.'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Uses the refresh_token cookie to issue a new access token. Rate limited to 0.5 requests per second with a burst of 5.
      operationId: refreshToken
      responses:
        "200":
          description: Token refreshed
          headers:
            Set-Cookie:
              description: refresh_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or missing refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Always returns 200 to prevent email enumeration. Rate limited to 0.5 requests per second with a burst of 5.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password with token
      description: Rate limited to 0.5 requests per second with a burst of 5.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token, or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/confirm-email:
    post:
      tags: [Authentication]
      summary: Confirm email address
      description: Verifies the confirmation token from the email link and activates the account.
      operationId: confirmEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmEmailRequest"
      responses:
        "200":
          description: Email confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/resend-confirmation:
    post:
      tags: [Authentication]
      summary: Resend confirmation email
      description: |-
        When an email backend is configured, returns 200 to prevent email enumeration; if the email exists and is unverified, a new confirmation email is sent (24-hour expiry).

        When no email backend is configured the endpoint returns 400 with `email backend not configured` because new accounts auto-verify and resend has no effect.
      operationId: resendConfirmation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendConfirmationRequest"
      responses:
        "200":
          description: Confirmation email sent if account exists and is unverified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error or no email backend configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/user:
    get:
      tags: [User]
      summary: Get current user profile
      description: Rate limited to 2 requests per second with a burst of 10.
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [User]
      summary: Update current user profile
      description: Rate limited to 2 requests per second with a burst of 10.
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong current password or unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/settings/notifications:
    get:
      tags: [Settings]
      summary: Get notification preferences
      description: Returns the current user's account-wide notification preferences.
      operationId: getNotificationPreferences
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Settings]
      summary: Update notification preferences
      description: Sets the account-wide notification preference for view and comment notifications.
      operationId: updateNotificationPreferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesRequest"
      responses:
        "204":
          description: Preferences updated
        "400":
          description: Invalid notification mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/settings/integrations:
    get:
      tags: [Integrations]
      operationId: listIntegrations
      summary: List configured integrations
      description: Returns all integrations for the authenticated user. Token fields are masked.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of integrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntegrationConfig"
        "401":
          description: Unauthorized
...