# internal/server/server.go
package server

import (
	"context"
	"fmt"
	"io/fs"
	"log"
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/sendrec/sendrec/internal/auth"
	"github.com/sendrec/sendrec/internal/billing"
	"github.com/sendrec/sendrec/internal/capture"
	"github.com/sendrec/sendrec/internal/database"
	"github.com/sendrec/sendrec/internal/docs"
	"github.com/sendrec/sendrec/internal/geoip"
	"github.com/sendrec/sendrec/internal/integration"
	"github.com/sendrec/sendrec/internal/organization"
	"github.com/sendrec/sendrec/internal/ratelimit"
	"github.com/sendrec/sendrec/internal/scim"
	"github.com/sendrec/sendrec/internal/sso"
	"github.com/sendrec/sendrec/internal/video"
	"github.com/sendrec/sendrec/internal/webhook"
)

type Pinger interface {
	Ping(ctx context.Context) error
}

type Config struct {
	Version                 string
	DB                      database.DBTX
	Pinger                  Pinger
	Storage                 video.ObjectStorage
	WebFS                   fs.FS
	JWTSecret               string
	BaseURL                 string
	MaxUploadBytes          int64
	MaxVideosPerMonth       int
	MaxVideoDurationSeconds int
	MaxPlaylists            int
	S3PublicEndpoint        string
	EnableDocs              bool
	BrandingEnabled         bool
	AiEnabled               bool
	TranscriptionEnabled    bool
	NoiseReductionFilter    string
	AllowedFrameAncestors   string
	AnalyticsScript         string
	// CaptureTokenSecret enables the cross-origin capture hand-off (a MajorGTM
	// fork addition). Empty means the /capture route is not registered at all —
	// a stock deployment has no parent application to trust.
	CaptureTokenSecret        string
	CaptureOrgMap             string
	EmailSender               auth.EmailSender
	CommentNotifier           video.CommentNotifier
	ViewNotifier              video.ViewNotifier
	SlackNotifier             video.SlackNotifier
	WebhookClient             *webhook.Client
	CreemAPIKey               string
	CreemWebhookSecret        string
	CreemProProductID         string
	CreemOrgProProductID      string
	CreemBusinessProductID    string
	CreemOrgBusinessProductID string
	RegistrationEnabled       bool
	PlanBadgeEnabled          bool
	GeoIPDBPath               string
	GoogleClientID            string
	GoogleClientSecret        string
	GoogleAllowedDomains      []string
	MicrosoftClientID         string
	MicrosoftClientSecret     string
	GitHubSSOClientID         string
	GitHubSSOClientSecret     string
}

type Server struct {
	router              chi.Router
	version             string
	pinger              Pinger
	authHandler         *auth.Handler
	videoHandler        *video.Handler
	orgHandler          *organization.Handler
	integrationHandler  *integration.Handler
	ssoHandler          *sso.Handler
	scimHandler         *scim.Handler
	captureHandler      *capture.Handler
	db                  database.DBTX
	billingHandlers     *billing.Handlers
	webFS               fs.FS
	enableDocs          bool
	registrationEnabled bool
	planBadgeEnabled    bool
	analyticsScript     string
}

func New(cfg Config) *Server {
	r := chi.NewRouter()
	r.Use(slogMiddleware)
	r.Use(middleware.Recoverer)
	r.Use(securityHeaders(SecurityConfig{
		BaseURL:               cfg.BaseURL,
		StorageEndpoint:       cfg.S3PublicEndpoint,
		AllowedFrameAncestors: cfg.AllowedFrameAncestors,
	}))

	s := &Server{router: r, version: cfg.Version, pinger: cfg.Pinger, db: cfg.DB, webFS: cfg.WebFS, enableDocs: cfg.EnableDocs, registrationEnabled: cfg.RegistrationEnabled, planBadgeEnabled: cfg.PlanBadgeEnabled, analyticsScript: cfg.AnalyticsScript}

	if cfg.DB != nil {
		jwtSecret := cfg.JWTSecret
		if jwtSecret == "" {
			log.Fatal("JWT_SECRET is required; set the environment variable")
		}

		baseURL := cfg.BaseURL
		if baseURL == "" {
			baseURL = "http://localhost:8080"
		}

		secureCookies := strings.HasPrefix(baseURL, "https://")
		s.authHandler = auth.NewHandler(cfg.DB, jwtSecret, secureCookies)
		s.authHandler.SetRegistrationEnabled(cfg.RegistrationEnabled)
		// Fork addition. Constructed only when a secret is configured, so the
		// route simply does not exist on a stock deployment.
		if cfg.CaptureTokenSecret != "" {
			s.captureHandler = capture.NewHandlerWithOrgMap(cfg.DB, jwtSecret, cfg.CaptureTokenSecret, cfg.AllowedFrameAncestors, cfg.CaptureOrgMap)
		}
		if cfg.EmailSender != nil {
			s.authHandler.SetEmailSender(cfg.EmailSender, baseURL)
		}
		s.videoHandler = video.NewHandler(cfg.DB, cfg.Storage, baseURL, cfg.MaxUploadBytes, cfg.MaxVideosPerMonth, cfg.MaxVideoDurationSeconds, cfg.MaxPlaylists, jwtSecret, secureCookies)
		if cfg.CommentNotifier != nil {
			s.videoHandler.SetCommentNotifier(cfg.CommentNotifier)
		}
		if cfg.ViewNotifier != nil {
			s.videoHandler.SetViewNotifier(cfg.ViewNotifier)
		}
		if cfg.BrandingEnabled {
			s.videoHandler.SetBrandingEnabled(true)
		}
		if cfg.AnalyticsScript != "" {
			s.videoHandler.SetAnalyticsScript(cfg.AnalyticsScript)
		}
		if cfg.AiEnabled {
			s.videoHandler.SetAIEnabled(true)
		}
		if cfg.TranscriptionEnabled {
			s.videoHandler.SetTranscriptionEnabled(true)
		}
		if cfg.NoiseReductionFilter != "" {
			s.videoHandler.SetNoiseReductionFilter(cfg.NoiseReductionFilter)
		}
		if cfg.SlackNotifier != nil {
			s.videoHandler.SetSlackNotifier(cfg.SlackNotifier)
		}
		if cfg.WebhookClient != nil {
			s.videoHandler.SetWebhookClient(cfg.WebhookClient)
		}
		if cfg.GeoIPDBPath != "" {
			geoResolver, err := geoip.New(cfg.GeoIPDBPath)
			if err == nil {
				s.videoHandler.SetGeoResolver(geoResolver)
			}
		}

		if cfg.CreemAPIKey != "" {
			creemClient := billing.New(cfg.CreemAPIKey, "")
			s.billingHandlers = billing.NewHandlers(cfg.DB, creemClient, baseURL, cfg.CreemProProductID, cfg.CreemOrgProProductID, cfg.CreemBusinessProductID, cfg.CreemOrgBusinessProductID, cfg.CreemWebhookSecret)
		}

		s.orgHandler = organization.NewHandler(cfg.DB, baseURL)
		if sender, ok := cfg.EmailSender.(organization.EmailSender); ok {
			s.orgHandler.SetEmailSender(sender)
		}

		s.integrationHandler = integration.NewHandler(cfg.DB, integration.DeriveKey(jwtSecret), baseURL)

		encKey := integration.DeriveKey(jwtSecret)
		s.ssoHandler = sso.NewHandler(cfg.DB, jwtSecret, baseURL, secureCookies, encKey)
		s.scimHandler = scim.NewHandler(cfg.DB, baseURL)

		if cfg.GoogleClientID != "" {
			googleProvider, err := sso.NewOIDCProvider(context.Background(), sso.OIDCConfig{
				IssuerURL:           "https://accounts.google.com",
				ClientID:            cfg.GoogleClientID,
				ClientSecret:        cfg.GoogleClientSecret,
				RedirectURL:         baseURL + "/api/auth/sso/google/callback",
				AllowedEmailDomains: cfg.GoogleAllowedDomains,
			})
			if err == nil {
				s.ssoHandler.RegisterProvider("google", googleProvider)
			}
		}
		if cfg.MicrosoftClientID != "" {
			msProvider, err := sso.NewOIDCProvider(context.Background(), sso.OIDCConfig{
				IssuerURL:    "https://login.microsoftonline.com/common/v2.0",
				ClientID:     cfg.MicrosoftClientID,
				ClientSecret: cfg.MicrosoftClientSecret,
				RedirectURL:  baseURL + "/api/auth/sso/microsoft/callback",
			})
			if err == nil {
				s.ssoHandler.RegisterProvider("microsoft", msProvider)
			}
		}
		if cfg.GitHubSSOClientID != "" {
			ghProvider := sso.NewGitHubProvider(sso.GitHubConfig{
				ClientID:     cfg.GitHubSSOClientID,
				ClientSecret: cfg.GitHubSSOClientSecret,
				RedirectURL:  baseURL + "/api/auth/sso/github/callback",
			})
			s.ssoHandler.RegisterProvider("github", ghProvider)
		}
	}

	s.routes()
	return s
}

func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	s.router.ServeHTTP(w, r)
}

func apiKeyOrJWTMiddleware(db database.DBTX, jwtMiddleware func(http.Handler) http.Handler) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			authHeader := r.Header.Get("Authorization")
			token, found := strings.CutPrefix(authHeader, "Bearer ")
			if found {
				userID, err := auth.LookupAPIKey(r.Context(), db, token)
				if err == nil {
					ctx := auth.ContextWithUserID(r.Context(), userID)
					next.ServeHTTP(w, r.WithContext(ctx))
					return
				}
			}
			jwtMiddleware(next).ServeHTTP(w, r)
		})
	}
}

func maxBodySize(maxBytes int64) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			if r.Body != nil {
				r.Body = http.MaxBytesReader(w, r.Body, maxBytes)
			}
			next.ServeHTTP(w, r)
		})
	}
}

func (s *Server) routes() {
	s.router.Get("/api/health", s.handleHealth)
	s.router.Get("/robots.txt", s.handleRobotsTxt)
	if s.enableDocs {
		s.router.Get("/api/docs", docs.HandleDocs)
		s.router.Get("/api/docs/openapi.yaml", docs.HandleSpec)
	}

	// The capture hand-off (fork addition). Its own limiter rather than the
	// login one: a person opening the recorder should not be turned away because
	// someone behind the same address fumbled a password five times, and the
	// token is HMAC-protected — this bounds abuse, it is not the defence.
	if s.captureHandler != nil {
		captureLimiter := ratelimit.NewLimiter(1, 10)
		s.router.With(captureLimiter.Middleware).Get("/capture", s.captureHandler.Redeem)
	}

	if s.authHandler != nil {
		authLimiter := ratelimit.NewLimiter(0.5, 5)
		s.router.Route("/api/auth", func(r chi.Router) {
			r.Use(authLimiter.Middleware)
			r.Use(maxBodySize(64 * 1024))
			r.Post("/register", s.authHandler.Register)
			r.Post("/login", s.authHandler.Login)
			r.Post("/refresh", s.authHandler.Refresh)
			r.Post("/logout", s.authHandler.Logout)
			r.Post("/forgot-password", s.authHandler.ForgotPassword)
			r.Post("/reset-password", s.authHandler.ResetPassword)
			r.Post("/confirm-email", s.authHandler.ConfirmEmail)
			r.Post("/resend-confirmation", s.authHandler.ResendConfirmation)
		})

		if s.ssoHandler != nil {
			s.router.Route("/api/auth/sso", func(r chi.Router) {
				r.Use(authLimiter.Middleware)
				r.Get("/providers", s.ssoHandler.Providers)
				r.Get("/org", s.ssoHandler.InitiateOrgSSO)
				r.Get("/org/callback", s.ssoHandler.OrgCallback)
				r.Get("/{provider}", s.ssoHandler.Initiate)
				r.Get("/{provider}/callback", s.ssoHandler.Callback)
			})

			// SAML routes (outside /api/auth/sso to avoid {provider} conflict)
			s.router.Route("/api/auth/saml", func(r chi.Router) {
				r.Use(authLimiter.Middleware)
				r.Post("/{orgId}/acs", s.ssoHandler.OrgSAMLCallback)
				r.Get("/{orgId}/metadata", s.ssoHandler.SPMetadata)
			})
		}

		// SCIM provisioning routes
		if s.scimHandler != nil {
			s.router.Route("/api/organizations/{orgId}/scim/v2", func(r chi.Router) {
				r.Use(scim.BearerAuth(s.db))
				r.Get("/ServiceProviderConfig", s.scimHandler.ServiceProviderConfig)
				r.Get("/Schemas", s.scimHandler.Schemas)
				r.Get("/Users", s.scimHandler.ListUsers)
				r.Get("/Users/{id}", s.scimHandler.GetUser)
				r.Post("/Users", s.scimHandler.CreateUser)
				r.Put("/Users/{id}", s.scimHandler.ReplaceUser)
				r.Patch("/Users/{id}", s.scimHandler.PatchUser)
				r.Delete("/Users/{id}", s.scimHandler.DeleteUser)
			})
		}
	}

	if s.authHandler != nil {
		s.router.Route("/api/user", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(maxBodySize(64 * 1024))
			r.Get("/", s.authHandler.GetUser)
			r.Patch("/", s.authHandler.UpdateUser)
			if s.ssoHandler != nil {
				r.Get("/identities", s.ssoHandler.ListIdentities)
				r.Delete("/identities/{provider}", s.ssoHandler.UnlinkIdentity)
			}
		})
	}

	if s.orgHandler != nil {
		s.router.Route("/api/organizations", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(maxBodySize(64 * 1024))
			r.Get("/", s.orgHandler.List)
			r.Post("/", s.orgHandler.Create)
			r.Route("/{orgId}", func(r chi.Router) {
				r.Get("/", s.orgHandler.Get)
				r.Patch("/", s.orgHandler.Update)
				r.Delete("/", s.orgHandler.Delete)
				r.Get("/members", s.orgHandler.ListMembers)
				r.Delete("/members/{userId}", s.orgHandler.RemoveMember)
				r.Patch("/members/{userId}", s.orgHandler.UpdateMemberRole)
				r.Post("/invites", s.orgHandler.SendInvite)
				r.Get("/invites", s.orgHandler.ListInvites)
				r.Delete("/invites/{inviteId}", s.orgHandler.RevokeInvite)
				if s.billingHandlers != nil {
					r.Route("/billing", func(r chi.Router) {
						r.Get("/", s.billingHandlers.GetOrgBilling)
						r.Post("/checkout", s.billingHandlers.CreateOrgCheckout)
						r.Delete("/", s.billingHandlers.CancelOrgSubscription)
					})
				}
				if s.ssoHandler != nil {
					r.Get("/sso", s.ssoHandler.GetConfig)
					r.Put("/sso", s.ssoHandler.SaveConfig)
					r.Delete("/sso", s.ssoHandler.DeleteConfig)
					r.Get("/scim-token", s.ssoHandler.GetSCIMToken)
					r.Post("/scim-token", s.ssoHandler.GenerateSCIMToken)
					r.Delete("/scim-token", s.ssoHandler.RevokeSCIMToken)
				}
			})
		})

		s.router.Route("/api/invites", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(maxBodySize(64 * 1024))
			r.Post("/accept", s.orgHandler.AcceptInvite)
		})
	}

	if s.videoHandler != nil {
		s.router.Route("/api/settings", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(maxBodySize(64 * 1024))
			r.Get("/notifications", s.videoHandler.GetNotificationPreferences)
			r.Put("/notifications", s.videoHandler.PutNotificationPreferences)
			r.Post("/notifications/test-slack", s.videoHandler.TestSlackWebhook)
			r.Post("/notifications/test-webhook", s.videoHandler.TestWebhook)
			r.Post("/notifications/regenerate-webhook-secret", s.videoHandler.RegenerateWebhookSecret)
			r.Get("/notifications/webhook-deliveries", s.videoHandler.ListWebhookDeliveries)
			r.Get("/branding", s.videoHandler.GetBrandingSettings)
			r.Put("/branding", s.videoHandler.PutBrandingSettings)
			r.Post("/branding/logo", s.videoHandler.UploadBrandingLogo)
			r.Delete("/branding/logo", s.videoHandler.DeleteBrandingLogo)
			r.Post("/api-keys", auth.GenerateAPIKey(s.db))
			r.Get("/api-keys", auth.ListAPIKeys(s.db))
			r.Delete("/api-keys/{id}", auth.DeleteAPIKey(s.db))
			if s.billingHandlers != nil {
				r.Get("/billing", s.billingHandlers.GetBilling)
				r.Post("/billing/checkout", s.billingHandlers.CreateCheckout)
				r.Post("/billing/cancel", s.billingHandlers.CancelSubscription)
			}
			if s.integrationHandler != nil {
				r.Get("/integrations", s.integrationHandler.List)
				r.Put("/integrations/{provider}", s.integrationHandler.Save)
				r.Delete("/integrations/{provider}", s.integrationHandler.Delete)
				r.Post("/integrations/{provider}/test", s.integrationHandler.Test)
			}
		})

		videoLimiter := ratelimit.NewLimiter(2, 10)
		s.router.Route("/api/videos", func(r chi.Router) {
			r.Use(videoLimiter.Middleware)
			r.Use(maxBodySize(64 * 1024))
			// List accepts API key OR JWT
			r.With(apiKeyOrJWTMiddleware(s.db, s.authHandler.Middleware), organization.Middleware(s.db)).Get("/", s.videoHandler.List)
			// All other endpoints require JWT
			r.Group(func(r chi.Router) {
				r.Use(s.authHandler.Middleware)
				r.Use(organization.Middleware(s.db))

				// Read-only routes (viewer allowed)
				r.Get("/limits", s.videoHandler.Limits)
				r.Get("/{id}/download", s.videoHandler.Download)
				r.Get("/{id}/comments", s.videoHandler.ListOwnerComments)
				r.Get("/{id}/analytics", s.videoHandler.Analytics)
				r.Get("/{id}/analytics/export", s.videoHandler.AnalyticsExport)
				r.Get("/{id}/branding", s.videoHandler.GetVideoBranding)

				// Write routes (viewer blocked)
				r.Group(func(r chi.Router) {
					r.Use(organization.RequireWriter)
					r.Post("/", s.videoHandler.Create)
					r.Post("/upload", s.videoHandler.Upload)
					r.Post("/batch/delete", s.videoHandler.BatchDelete)
					r.Post("/batch/folder", s.videoHandler.BatchSetFolder)
					r.Post("/batch/tags", s.videoHandler.BatchSetTags)
					r.Patch("/{id}", s.videoHandler.Update)
					r.Delete("/{id}", s.videoHandler.Delete)
					r.Post("/{id}/extend", s.videoHandler.Extend)
					r.Post("/{id}/trim", s.videoHandler.Trim)
					r.Post("/{id}/retranscribe", s.videoHandler.Retranscribe)
					r.Put("/{id}/password", s.videoHandler.SetPassword)
					r.Put("/{id}/comment-mode", s.videoHandler.SetCommentMode)
					r.Delete("/{id}/comments/{commentId}", s.videoHandler.DeleteComment)
					r.Put("/{id}/notifications", s.videoHandler.SetVideoNotification)
					r.Put("/{id}/download-enabled", s.videoHandler.SetDownloadEnabled)
					r.Put("/{id}/link-expiry", s.videoHandler.SetLinkExpiry)
					r.Put("/{id}/branding", s.videoHandler.SetVideoBranding)
					r.Post("/{id}/thumbnail", s.videoHandler.UploadThumbnail)
					r.Delete("/{id}/thumbnail", s.videoHandler.ResetThumbnail)
					r.Put("/{id}/cta", s.videoHandler.SetCTA)
					r.Put("/{id}/email-gate", s.videoHandler.SetEmailGate)
					r.Post("/{id}/summarize", s.videoHandler.Summarize)
					r.Post("/{id}/generate-document", s.videoHandler.GenerateDocument)
					r.Put("/{id}/folder", s.videoHandler.SetVideoFolder)
					r.Put("/{id}/tags", s.videoHandler.SetVideoTags)
					r.Post("/{id}/remove-segments", s.videoHandler.RemoveSegments)
					r.Post("/{id}/detect-silence", s.videoHandler.DetectSilence)
					r.Put("/{id}/dismiss-title", s.videoHandler.DismissTitle)
					r.Put("/{id}/pin", s.videoHandler.TogglePin)
					r.Post("/{id}/transfer", s.videoHandler.Transfer)
					if s.integrationHandler != nil {
						r.Post("/{id}/create-issue", s.integrationHandler.CreateIssue)
					}
				})
			})
		})

		// Separate mounted subrouter: GET and POST /api/videos/{id}/transcript
		// live together here so neither is shadowed by the other (a Route()
		// subtree mounted at a path that collides with a sibling route
		// registered inside the /api/videos group wins the routing decision
		// for ALL methods at that path, 405-ing the sibling). Mounting both
		// verbs in one subrouter also lets the POST route use a larger
		// body-size cap without the /api/videos group's 64KB maxBodySize
		// double-wrapping it (chi's With() adds middleware on top of,
		// not instead of, the enclosing group's middleware).
		s.router.Route("/api/videos/{id}/transcript", func(r chi.Router) {
			r.Use(videoLimiter.Middleware)
			r.Use(s.authHandler.Middleware)
			r.Use(organization.Middleware(s.db))
			r.Get("/", s.videoHandler.GetTranscript)
			r.With(maxBodySize(video.MaxTranscriptUploadBytes+1024), organization.RequireWriter).
				Post("/", s.videoHandler.UploadTranscript)
		})

		s.router.Route("/api/analytics", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(organization.Middleware(s.db))
			r.Get("/dashboard", s.videoHandler.AnalyticsDashboard)
			r.Get("/dashboard/export", s.videoHandler.DashboardExport)
		})

		s.router.Route("/api/folders", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(organization.Middleware(s.db))
			r.Use(maxBodySize(64 * 1024))
			r.Get("/", s.videoHandler.ListFolders)
			r.Group(func(r chi.Router) {
				r.Use(organization.RequireWriter)
				r.Post("/", s.videoHandler.CreateFolder)
				r.Put("/{id}", s.videoHandler.UpdateFolder)
				r.Delete("/{id}", s.videoHandler.DeleteFolder)
			})
		})

		s.router.Route("/api/tags", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(organization.Middleware(s.db))
			r.Use(maxBodySize(64 * 1024))
			r.Get("/", s.videoHandler.ListTags)
			r.Group(func(r chi.Router) {
				r.Use(organization.RequireWriter)
				r.Post("/", s.videoHandler.CreateTag)
				r.Put("/{id}", s.videoHandler.UpdateTag)
				r.Delete("/{id}", s.videoHandler.DeleteTag)
			})
		})

		s.router.Route("/api/playlists", func(r chi.Router) {
			r.Use(s.authHandler.Middleware)
			r.Use(maxBodySize(64 * 1024))
			r.Get("/", s.videoHandler.ListPlaylists)
			r.Get("/{id}", s.videoHandler.GetPlaylist)
			r.Group(func(r chi.Router) {
				r.Use(organization.RequireWriter)
				r.Post("/", s.videoHandler.CreatePlaylist)
				r.Patch("/{id}", s.videoHandler.UpdatePlaylist)
				r.Delete("/{id}", s.videoHandler.DeletePlaylist)
				r.Post("/{id}/videos", s.videoHandler.AddPlaylistVideos)
				r.Delete("/{id}/videos/{videoId}", s.videoHandler.RemovePlaylistVideo)
				r.Patch("/{id}/videos/reorder", s.videoHandler.ReorderPlaylistVideos)
			})
		})

		watchAuthLimiter := ratelimit.NewLimiter(0.5, 5)
		commentLimiter := ratelimit.NewLimiter(0.2, 3)
		commentReadLimiter := ratelimit.NewLimiter(5, 20)
		s.router.Get("/api/watch/{shareToken}", s.videoHandler.Watch)
		s.router.Get("/api/watch/{shareToken}/download", s.videoHandler.WatchDownload)
		s.router.With(watchAuthLimiter.Middleware, maxBodySize(64*1024)).Post("/api/watch/{shareToken}/verify", s.videoHandler.VerifyWatchPassword)
		s.router.With(commentReadLimiter.Middleware).Get("/api/watch/{shareToken}/comments", s.videoHandler.ListWatchComments)
		s.router.With(commentLimiter.Middleware, maxBodySize(64*1024)).Post("/api/watch/{shareToken}/comments", s.videoHandler.PostWatchComment)
		s.router.With(watchAuthLimiter.Middleware, maxBodySize(64*1024)).Post("/api/watch/{shareToken}/identify", s.videoHandler.IdentifyViewer)
		s.router.Post("/api/watch/{shareToken}/cta-click", s.videoHandler.RecordCTAClick)
		s.router.Post("/api/watch/{shareToken}/milestone", s.videoHandler.RecordMilestone)
		s.router.Post("/api/watch/{shareToken}/segments", s.videoHandler.RecordSegments)
		s.router.Get("/api/watch/{shareToken}/thumbnail", s.videoHandler.WatchThumbnail)
		s.router.Get("/api/videos/{shareToken}/oembed", s.videoHandler.OEmbed)
		s.router.Get("/watch/{shareToken}", s.videoHandler.WatchPage)
		s.router.Get("/embed/{shareToken}", s.videoHandler.EmbedPage)

		s.router.Get("/watch/playlist/{shareToken}", s.videoHandler.PlaylistWatchPage)
		s.router.Get("/embed/playlist/{shareToken}", s.videoHandler.PlaylistEmbedPage)
		s.router.With(watchAuthLimiter.Middleware, maxBodySize(64*1024)).Post("/api/watch/playlist/{shareToken}/verify", s.videoHandler.VerifyPlaylistWatchPassword)
		s.router.With(watchAuthLimiter.Middleware, maxBodySize(64*1024)).Post("/api/watch/playlist/{shareToken}/identify", s.videoHandler.IdentifyPlaylistViewer)

		if s.billingHandlers != nil {
			s.router.Post("/api/webhooks/creem", s.billingHandlers.Webhook)
		}
	}

	if s.webFS != nil {
		spa := newSPAFileServer(s.webFS, s.analyticsScript)
		s.router.NotFound(spa.ServeHTTP)
	}
}

func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	if s.pinger != nil {
		if err := s.pinger.Ping(r.Context()); err != nil {
			w.WriteHeader(http.StatusServiceUnavailable)
			_, _ = fmt.Fprintf(w, `{"status":"unhealthy","version":%q,"error":"database unreachable"}`, s.version)
			return
		}
	}
	_, _ = fmt.Fprintf(w, `{"status":"ok","version":%q,"registrationEnabled":%t,"planBadgeEnabled":%t}`, s.version, s.registrationEnabled, s.planBadgeEnabled)
}

func (s *Server) handleRobotsTxt(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
	_, _ = w.Write([]byte("User-agent: *\nAllow: /watch/\nAllow: /embed/\nDisallow: /\n"))
}